chore: Add aliases and shared folders for renaming (#1107)
