+++ client/configuration/webpack/webpack.config.shared.js
@@ -34,6 +34,10 @@ module.exports = {
   },
   resolve: {
     extensions: [".ts", ".tsx", ".js", ".jsx", "..."],
+    modules: [path.resolve(__dirname, "../../src"), "node_modules"],
+    alias: {
+      "~/globals": path.resolve(__dirname, "../../src/globals"),
+    },
   },
   module: {
     rules: [
