+++ client/src/common/components/BrushableHistogram/components/ErrorLoading/ErrorLoading.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import * as globals from "../../globals";
+import * as globals from "~/globals";
 
 // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-explicit-any -- - FIXME: disabled temporarily on migrate to TS.
 const ErrorLoading = ({ displayName, zebra }: any) => (
