+++ client/src/common/components/BrushableHistogram/components/HistogramFooter/HistogramFooter.tsx
