+++ client/src/common/components/BrushableHistogram/components/HistogramHeader/HistogramHeader.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback } from "react";
 import { Button, ButtonGroup, Icon, Tooltip } from "@blueprintjs/core";
 import { IconNames } from "@blueprintjs/icons";
-import * as globals from "../../globals";
+import * as globals from "~/globals";
 
 const HistogramHeader = React.memo(
   ({
