+++ client/src/common/components/BrushableHistogram/utils.ts
