+++ client/src/components/BottomBanner/BottomBanner.tsx
@@ -1,5 +1,6 @@
 import React, { memo } from "react";
 import { connect } from "react-redux";
+import { AppDispatch, RootState } from "reducers";
 import {
   BOTTOM_BANNER_ID,
   StyledBanner,
@@ -10,7 +11,6 @@ import {
   BOTTOM_BANNER_SURVEY_LINK_TEXT,
   BOTTOM_BANNER_SURVEY_TEXT,
 } from "./constants";
-import { AppDispatch, RootState } from "../../reducers";
 
 export interface BottomBannerProps {
   surveyLink: string;
